Basic framework of the game is implemented
Files need to be refactored and the comments need to be cleaned up. A few bugs are still in the game, and there should be some menu refinement, but it is functional and will serve as a foundation for the structure before we start to port it over into our game engine.